part1, with some issues

+++ AoC2024/src/main/java/org/example/day7/FileProcessor.java
@@ -8,20 +8,20 @@ import java.util.stream.Collectors;
 
 public class FileProcessor {
 
-    public Map<Integer, List<Integer>> readFile(String filePath){
-        Map<Integer, List<Integer>> result = new HashMap<>();
+    public Map<Long, List<Long>> readFile(String filePath){
+        Map<Long, List<Long>> result = new HashMap<>();
         try(BufferedReader br = new BufferedReader(new FileReader(filePath))){
             String line;
             while((line = br.readLine()) != null){
                 int n = line.length();
                 String[] numberGroup = line.split(":");
                 String[] numbers = numberGroup[1].trim().split("\\s+");
-                List<Integer> list = new ArrayList<>();
+                List<Long> list = new ArrayList<>();
                 for (String number : numbers) {
-                    int i = Integer.parseInt(number);
+                    long i = Long.parseLong(number);
                     list.add(i);
                 }
-                result.put(Integer.parseInt(numberGroup[0]), list);
+                result.put(Long.parseLong(numberGroup[0]), list);
             }
         } catch(IOException e){
             e.printStackTrace();
